# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Maiko API Documentation",
        "contact": {
            "email": "juancho637@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/maiko/public/api/v1"
        }
    ],
    "paths": {
...
        "/corrosions/{corrosion}/files": {
            "get": {
                "tags": [
                    "Corrosiones"
                ],
                "summary": "Listado de archivos asociadas a una corrosión",
                "operationId": "App\\Http\\Controllers\\Api\\Corrosion\\CorrosionFileController::index",
                "parameters": [
                    {
                        "name": "corrosion",
                        "in": "path",
                        "description": "Id de la corrosión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra el listado de archivos asociadas a una corrosión.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Id no encontrado."
                    },
                    "401": {
                        "description": "Usuario no autorizado."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Corrosiones"
                ],
                "summary": "Almacena un archivo asociado a una corrosión",
                "operationId": "App\\Http\\Controllers\\Api\\Corrosion\\CorrosionFileController::store",
                "parameters": [
                    {
                        "name": "corrosion",
                        "in": "path",
                        "description": "Id de la corrosión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Almacena un nuevo archivo asociado a una corrosión.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Usuario no autorizado."
                    },
                    "409": {
                        "description": "Error en validaciones de negocio."
                    },
                    "422": {
                        "description": "Entidad no procesable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/corrosions/{corrosion}/files/{file}": {
            "get": {
                "tags": [
                    "Corrosiones"
                ],
                "summary": "Muestra un archivo asociado a una corrosión",
                "operationId": "App\\Http\\Controllers\\Api\\Corrosion\\CorrosionFileController::show",
                "parameters": [
                    {
                        "name": "corrosion",
                        "in": "path",
                        "description": "Id de la corrosión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "file",
                        "in": "path",
                        "description": "Id del archivo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Muestra un archivo asociado a una corrosión."
                    },
                    "401": {
                        "description": "Usuario no autorizado."
                    },
                    "409": {
                        "description": "Error en validaciones de negocio."
                    },
                    "422": {
                        "description": "Entidad no procesable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Corrosiones"
                ],
                "summary": "Actualiza un archivo asociado a una corrosión",
                "operationId": "App\\Http\\Controllers\\Api\\Corrosion\\CorrosionFileController::update",
                "parameters": [
                    {
                        "name": "corrosion",
                        "in": "path",
                        "description": "Id de la corrosión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "file",
                        "in": "path",
                        "description": "Id del archivo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualiza un archivo asociado a una corrosión.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Usuario no autorizado."
                    },
                    "409": {
                        "description": "Error en validaciones de negocio."
                    },
                    "422": {
                        "description": "Entidad no procesable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Corrosiones"
                ],
                "summary": "Elimina un archivo asociado a una corrosión",
                "operationId": "App\\Http\\Controllers\\Api\\Corrosion\\CorrosionFileController::destroy",
                "parameters": [
                    {
                        "name": "corrosion",
                        "in": "path",
                        "description": "Id de la corrosión",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "file",
                        "in": "path",
                        "description": "Id del archivo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Elimina un archivo asociado a una corrosión.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Usuario no autorizado."
                    },
                    "409": {
                        "description": "Error en validaciones de negocio."
                    },
                    "422": {
                        "description": "Entidad no procesable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}